refactor(config): share config screen between routes
